#16373: Move conv specific ops from tensor_utils to conv2d.
