+++ tests/tt_eager/ops/test_conv_prepare_weights_and_biases.cpp
@@ -5,7 +5,6 @@
 #include "common/assert.hpp"
 #include "common/bfloat16.hpp"
 #include "ttnn/cpp/ttnn/tensor/host_buffer/functions.hpp"
-#include "ttnn/cpp/ttnn/tensor/tensor_utils.hpp"
 #include "ttnn/cpp/ttnn/tensor/types.hpp"
 #include "ttnn/tensor/host_buffer/functions.hpp"
 #include "ttnn/tensor/host_buffer/types.hpp"
@@ -14,7 +13,7 @@
 #include "ttnn/operations/creation.hpp"
 #include "ttnn/operations/functions.hpp"
 #include "ttnn/tensor/types.hpp"
-#include "ttnn/tensor/tensor_utils.hpp"
+#include "ttnn/cpp/ttnn/operations/conv/conv2d/prepare_conv2d_weights.hpp"
 
 static std::vector<std::vector<bfloat16>> ref_weight_in = {
     {
@@ -452,8 +451,8 @@ static void test_convert_conv_weight_tensor_to_tiled_layout_block_sharded() {
         for (auto j = 0; j < input_buffer.size(); j++) {
             input_buffer[j] = ref_weight_in[i][j];
         }
-        auto output_tensor =
-            convert_conv_weight_tensor_to_tiled_layout_block_sharded(input_tensor, shards[i], DataType::BFLOAT16);
+        auto output_tensor = ttnn::operations::conv::convert_conv_weight_tensor_to_tiled_layout_block_sharded(
+            input_tensor, shards[i], DataType::BFLOAT16);
         auto out_buffer = owned_buffer::get_as<bfloat16>(output_tensor);
 
         TT_FATAL(compare_out_with_ref(out_buffer, ref_weight_out[i]) == 0, "Error");
@@ -465,8 +464,8 @@ static void test_convert_conv_bias_tensor_to_tiled_layout_block_sharded() {
     for (auto i = 0; i < bias_tensor_shape.size(); i++) {
         auto input_tensor = ttnn::random::random(bias_tensor_shape[i], DataType::BFLOAT16).to(Layout::ROW_MAJOR).cpu();
         auto input_buffer = owned_buffer::get_as<bfloat16>(input_tensor);
-        auto output_tensor =
-            convert_conv_bias_tensor_to_tiled_layout_block_sharded(input_tensor, shards[i], DataType::BFLOAT16);
+        auto output_tensor = ttnn::operations::conv::convert_conv_bias_tensor_to_tiled_layout_block_sharded(
+            input_tensor, shards[i], DataType::BFLOAT16);
         auto out_buffer = owned_buffer::get_as<bfloat16>(output_tensor);
         /* Expected output should be same as input buffer except some padding*/
         TT_FATAL(compare_out_with_ref(out_buffer, input_buffer) == 0, "Error");
